+++ packages/desktop/src/actions/ActionsSize.ts
@@ -1,3 +1,0 @@
-type ActionsSize = 'accent' | 'normal' | 'minor'
-
-export default ActionsSize