+++ packages/desktop/src/button/Button.tsx
@@ -2,7 +2,7 @@ import React, {FunctionComponent, ReactNode, Fragment} from 'react'
 
 import {ButtonControl, Spinner, Pos, Flex, FlexItem, Typo, Btn} from '@qiwi/pijma-core'
 
-interface ButtonProps {
+export interface ButtonProps {
   onClick?: () => void
   onFocus?: () => void
   onBlur?: () => void
@@ -23,6 +23,11 @@ const buttonBackground: { [kind in ButtonProps['kind']]: string } = {
 
 const buttonHoverBackground: { [kind in ButtonProps['kind']]: string } = {
   brand: '#ff8200',
+  simple: '#fff padding-box',
+}
+
+const accentButtonBackground: { [kind in ButtonProps['kind']]: string } = {
+  brand: 'linear-gradient(to bottom, #ff9810, #ff8300)',
   simple: '#fff padding-box',
 }
 
@@ -71,7 +76,17 @@ const textSize: { [size in ButtonProps['size']]: number } = {
   minor: 3.5,
 }
 
-const Button: FunctionComponent<ButtonProps> = (props) => (
+const shadow: { [kind in ButtonProps['kind']]: string } = {
+  brand: '0 15px 50px -10px rgb(255, 206, 135)',
+  simple: '0 15px 50px -10px rgba(0, 0, 0, 0.15)',
+}
+
+const hoverShadow: { [kind in ButtonProps['kind']]: string } = {
+  brand: '0 25px 50px -10px rgb(255, 206, 135)',
+  simple: '0 25px 50px -10px rgba(0, 0, 0, 0.15)',
+}
+
+export const Button: FunctionComponent<ButtonProps> = (props) => (
   <ButtonControl
     onClick={props.onClick}
     onFocus={props.onFocus}
@@ -82,10 +97,11 @@ const Button: FunctionComponent<ButtonProps> = (props) => (
         width={!props.icon || props.text ? 1 : buttonSize[props.size]}
         height={buttonSize[props.size]}
         minWidth={buttonMinWith[props.size]}
-        bg={props.disabled ? '#e6e6e6' : renderProps.hover || renderProps.focus ? buttonHoverBackground[props.kind] : buttonBackground[props.kind]}
+        bg={props.disabled ? '#e6e6e6' : renderProps.hover || renderProps.focus ? buttonHoverBackground[props.kind] : props.size === 'accent' ? accentButtonBackground[props.kind] : buttonBackground[props.kind]}
         b={props.disabled ? 'none' : renderProps.hover || renderProps.focus ? buttonHoverBorder[props.kind] : buttonBorder[props.kind]}
         r={buttonRadius[props.size]}
-        transition="all 300ms cubic-bezier(0.4, 0.0, 0.2, 1)"
+        s={props.disabled ? 'none' : (renderProps.hover || renderProps.focus) && props.size === 'accent' ? hoverShadow[props.kind] : props.size === 'accent' ? shadow[props.kind] : 'none'}
+        transition="box-shadow 300ms cubic-bezier(0.4, 0.0, 0.2, 1)"
         onClick={renderProps.onClick}
         onFocus={renderProps.onFocus}
         onBlur={renderProps.onFocus}
@@ -147,7 +163,7 @@ const Button: FunctionComponent<ButtonProps> = (props) => (
                     ) : (
                       null
                     )}
-                    {props.text ? (
+                    {props.text || !props.icon ? (
                       <FlexItem
                         opacity={props.loading ? 0 : renderProps.hover || renderProps.active || renderProps.focus ? 0.9 : 1}
                         overflow="hidden"
@@ -178,5 +194,3 @@ const Button: FunctionComponent<ButtonProps> = (props) => (
     )}
   />
 )
-
-export default Button