+++ packages/mobile/src/actions/Actions.tsx
@@ -1,26 +1,30 @@
 import React, {FunctionComponent, ReactNode, Children} from 'react'
 
-import {styled, indent} from '@qiwi/pijma-core'
+import {Box} from '@qiwi/pijma-core'
 
-const ActionsList = styled('div')({
-  display: 'block',
-  width: '100%',
-})
+export interface ActionsProps {
+  size: 'accent' | 'normal' | 'minor'
+}
 
-const ActionsItem = styled('div')({
-  display: 'block',
-  width: '100%',
-  ...indent(20),
-})
-
-const Actions: FunctionComponent<{}> = (props) => (
-  <ActionsList>
-    {Children.map(props.children, (child: ReactNode, key: number) => (
-      <ActionsItem key={key}>
-        {child}
-      </ActionsItem>
-    ))}
-  </ActionsList>
-)
-
-export default Actions
+export const Actions: FunctionComponent<ActionsProps> = (props) => {
+  const elements = Children.toArray(props.children).filter(child => !!child)
+  if (elements.length === 0) {
+    return null
+  }
+  return (
+    <Box
+      width={1}
+      children={(
+        Children.map(elements, (child: ReactNode, key: number) => (
+          <Box
+            key={key}
+            width={1}
+            maxWidth={1}
+            mt={key === 0 ? undefined : 4}
+            children={child}
+          />
+        ))
+      )}
+    />
+  )
+}